+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 
 - [TD-7221] Highlight the headers of the template fields in Concepts download file
+- [TD-6468] Multi language upload and download business concepts
 
 ## [7.5.2] 2025-06-05